Solve review comments
