+++ Tests/Common/Securities/SecurityServiceTests.cs
@@ -19,6 +19,7 @@ using System.Linq;
 using Moq;
 using NUnit.Framework;
 using QuantConnect.Algorithm.CSharp;
+using QuantConnect.Configuration;
 using QuantConnect.Data;
 using QuantConnect.Data.Auxiliary;
 using QuantConnect.Data.Market;
@@ -26,6 +27,8 @@ using QuantConnect.Data.UniverseSelection;
 using QuantConnect.Interfaces;
 using QuantConnect.Securities;
 using QuantConnect.Tests.Engine.DataFeeds;
+using QuantConnect.ToolBox.RandomDataGenerator;
+using QuantConnect.Util;
 
 namespace QuantConnect.Tests.Common.Securities
 {
@@ -216,7 +219,7 @@ namespace QuantConnect.Tests.Common.Securities
             var equityOptionSecurity = (QuantConnect.Securities.Option.Option)_securityService.CreateSecurity(equityOption, configs, 1.0m);
 
             Assert.AreEqual(100, equityOptionSecurity.ContractMultiplier);
-            Assert.AreEqual(100,equityOptionSecurity.ContractUnitOfTrade);
+            Assert.AreEqual(100, equityOptionSecurity.ContractUnitOfTrade);
         }
 
         [Test]
@@ -276,6 +279,37 @@ namespace QuantConnect.Tests.Common.Securities
             Assert.AreEqual(equity.Subscriptions.First().Type, typeof(TradeBar));
             Assert.AreEqual(equity.Subscriptions.First().TickType, TickType.Trade);
             Assert.AreEqual(((QuantConnect.Securities.Equity.Equity)equity).PrimaryExchange, Exchange.NASDAQ);
+        }
+
+        [Test]
+        public void CreateSecurityDoesNotThrowWithNullAlgorithm()
+        {
+            var startDate = new DateTime(2024, 1, 1);
+            var securityManager = new SecurityManager(new TimeKeeper(startDate, new[] { TimeZones.Utc }));
+
+            var securityService = new SecurityService(
+                new CashBook(),
+                MarketHoursDatabase.FromDataFolder(),
+                SymbolPropertiesDatabase.FromDataFolder(),
+                new SecurityInitializerProvider(new FuncSecurityInitializer(security => { })),
+                RegisteredSecurityDataTypesProvider.Null,
+                new SecurityCacheProvider(
+                    new SecurityPortfolioManager(securityManager,
+                        new SecurityTransactionManager(null, securityManager),
+                        new AlgorithmSettings())),
+                new MapFilePrimaryExchangeProvider(
+                    Composer.Instance.GetExportedValueByTypeName<IMapFileProvider>(
+                        Config.Get("map-file-provider", "LocalDiskMapFileProvider"))),
+                algorithm: null
+            );
+
+            securityManager.SetSecurityService(securityService);
+
+            Assert.DoesNotThrow(() =>
+            {
+                var symbol = Symbol.Create("TEST", SecurityType.Equity, Market.USA);
+                var security = securityManager.CreateSecurity(symbol, new List<SubscriptionDataConfig>(), underlying: null);
+            });
         }
     }
 }
